feat: Introduce new CSV analysis, refactor PowerPoint generation, and add comprehensive unit tests.
