+++ _archive/test_nan.py
