+++ _archive/trashed/data_intelligence_fast_report_maker.py
