+++ pptx_builder/__init__.py
@@ -1,0 +1,3 @@
+# PowerPoint generation pipeline
+from .engine import generate_pptx, set_text_style, add_dataframe_as_table
+from .native_charts import add_native_line_chart, add_native_pie_chart
