+++ templates/index.html
@@ -57,7 +57,7 @@
 
     <label class="form-check" style="margin-bottom: 20px;">
       <input type="checkbox" name="solo_fecha" />
-      Solo fecha (sin hora)
+      Incluir hora en gráfico de conversación
     </label>
 
     <button type="submit" class="btn btn-primary btn-block btn-lg" id="btnSubmit">
