--- trashed/README.md
@@ -1,4 +1,0 @@
-# Powerpoint_automation
-
-A way to automate the creation of a powerpoint presentation using Python and the pptx library.
-See video: https://youtu.be/STUNieOfv1g
